+++ src/tx-builder/config/index.ts
@@ -11,6 +11,7 @@ import {
 } from '../types';
 
 export const DEFAULT_NULL_VALUE_ON_TX = BigNumber.from(0).toHexString();
+export const POLYGON_DEFAULT_GAS = BigNumber.from('1000000');
 export const DEFAULT_APPROVE_AMOUNT = constants.MaxUint256.toString();
 export const MAX_UINT_AMOUNT =
   '115792089237316195423570985008687907853269984665640564039457584007913129639935';
@@ -68,29 +69,50 @@ export const distinctStakingAddressesBetweenTokens: StakingConfigType = {
   [Stake.Aave]: {
     [Network.kovan]: {
       TOKEN_STAKING_ADDRESS: '0xf2fbf9A6710AfDa1c4AaB2E922DE9D69E0C97fd2',
+      STAKING_REWARD_TOKEN_ADDRESS:
+        '0xb597cd8d3217ea6477232f9217fa70837ff667af',
     },
     [Network.ropsten]: {
       TOKEN_STAKING_ADDRESS: '',
+      STAKING_REWARD_TOKEN_ADDRESS: '',
     },
     [Network.mainnet]: {
       TOKEN_STAKING_ADDRESS: '0x4da27a545c0c5b758a6ba100e3a049001de870f5',
+      STAKING_REWARD_TOKEN_ADDRESS:
+        '0x7fc66500c84a76ad7e9c93437bfc5ac33e2ddae9',
     },
     [Network.polygon]: {
       TOKEN_STAKING_ADDRESS: '',
+      STAKING_REWARD_TOKEN_ADDRESS: '',
+    },
+    [Network.mumbai]: {
+      TOKEN_STAKING_ADDRESS: '',
+      STAKING_REWARD_TOKEN_ADDRESS: '',
     },
   },
   [Stake.Balancer]: {
     [Network.kovan]: {
       TOKEN_STAKING_ADDRESS: '0xCe7021eDabaf82D28adBBea449Bc4dF70261F33E',
+      TOKEN_STAKING_ADDRESS: '0x31ce45Ab6E26C72c47C52c27498D460099545ef2',
+      STAKING_REWARD_TOKEN_ADDRESS:
+        '0xb597cd8d3217ea6477232f9217fa70837ff667af',
     },
     [Network.ropsten]: {
       TOKEN_STAKING_ADDRESS: '',
+      STAKING_REWARD_TOKEN_ADDRESS: '',
     },
     [Network.mainnet]: {
       TOKEN_STAKING_ADDRESS: '0xa1116930326D21fB917d5A27F1E9943A9595fb47',
+      STAKING_REWARD_TOKEN_ADDRESS:
+        '0x7fc66500c84a76ad7e9c93437bfc5ac33e2ddae9',
     },
     [Network.polygon]: {
       TOKEN_STAKING_ADDRESS: '',
+      STAKING_REWARD_TOKEN_ADDRESS: '',
+    },
+    [Network.mumbai]: {
+      TOKEN_STAKING_ADDRESS: '',
+      STAKING_REWARD_TOKEN_ADDRESS: '',
     },
   },
 };
@@ -140,16 +162,27 @@ export const enabledNetworksByService: EnabledNetworksType = {
   },
   claimStakingRewardsHelper: [Network.kovan],
   lendingPool: {
-    [Market.Proto]: [Network.kovan, Network.mainnet, Network.polygon],
+    [Market.Proto]: [
+      Network.kovan,
+      Network.mainnet,
+      Network.polygon,
+      Network.mumbai,
+    ],
     [Market.AMM]: [Network.kovan, Network.mainnet],
   },
   governance: [Network.kovan, Network.mainnet],
-  wethGateway: [Network.kovan, Network.mainnet, Network.polygon],
+  wethGateway: [
+    Network.kovan,
+    Network.mainnet,
+    Network.polygon,
+    Network.mumbai,
+  ],
   faucet: [Network.kovan],
   liquiditySwapAdapter: [Network.kovan, Network.mainnet],
   repayWithCollateralAdapter: [Network.kovan, Network.mainnet],
   aaveGovernanceV2: [Network.kovan, Network.mainnet],
   ltaMigrator: [Network.kovan, Network.mainnet],
+  incentivesController: [Network.polygon, Network.mumbai, Network.mainnet],
 };
 
 export * from './v2/addresses';