+++ src/tx-builder/services/v2/LendingPool.ts
@@ -1,4 +1,4 @@
-import { BigNumber, constants, utils } from 'ethers';
+import { constants, utils, BigNumberish, BytesLike } from 'ethers';
 import {
   API_ETH_MOCK_ADDRESS,
   commonContractAddressBetweenMarketsV2,
@@ -21,6 +21,7 @@ import {
   TokenMetadataType,
   transactionType,
   tStringDecimalUnits,
+  tEthereumAddress,
 } from '../../types';
 import { getTxValue, parseNumber } from '../../utils/parsings';
 import { LPValidator } from '../../validators/methodValidators';
@@ -42,9 +43,42 @@ import {
   IsPositiveAmount,
   IsPositiveOrMinusOneAmount,
 } from '../../validators/paramValidators';
-import LiquiditySwapAdapterInterface from '../../interfaces/LiquiditySwapAdapter';
+import LiquiditySwapAdapterInterface from '../../interfaces/LiquiditySwapAdapterParaswap';
 import RepayWithCollateralAdapterInterface from '../../interfaces/RepayWithCollateralAdapter';
 import BaseService from '../BaseService';
+import { augustusFromAmountOffsetFromCalldata } from '../LiquiditySwapAdapterParaswap';
+
+const buildParaSwapLiquiditySwapParams = (
+  assetToSwapTo: tEthereumAddress,
+  minAmountToReceive: BigNumberish,
+  swapAllBalanceOffset: BigNumberish,
+  swapCalldata: string | Buffer | BytesLike,
+  augustus: tEthereumAddress,
+  permitAmount: BigNumberish,
+  deadline: BigNumberish,
+  v: BigNumberish,
+  r: string | Buffer | BytesLike,
+  s: string | Buffer | BytesLike
+) => {
+  return utils.defaultAbiCoder.encode(
+    [
+      'address',
+      'uint256',
+      'uint256',
+      'bytes',
+      'address',
+      'tuple(uint256,uint256,uint8,bytes32,bytes32)',
+    ],
+    [
+      assetToSwapTo,
+      minAmountToReceive,
+      swapAllBalanceOffset,
+      swapCalldata,
+      augustus,
+      [permitAmount, deadline, v, r, s],
+    ]
+  );
+};
 
 export default class LendingPool
   extends BaseService<ILendingPool>
@@ -531,8 +565,9 @@ export default class LendingPool
     @IsEthAddress('fromAToken')
     @IsEthAddress('toAsset')
     @IsEthAddress('onBehalfOf')
+    @IsEthAddress('augustus')
     @IsPositiveAmount('fromAmount')
-    @IsPositiveAmount('toAmount')
+    @IsPositiveAmount('minToAmount')
     {
       user,
       flash,
@@ -540,13 +575,13 @@ export default class LendingPool
       fromAToken,
       toAsset,
       fromAmount,
-      toAmount,
-      maxSlippage,
+      minToAmount,
       permitSignature,
       swapAll,
       onBehalfOf,
       referralCode,
-      useEthPath,
+      augustus,
+      swapCallData,
     }: LPSwapCollateral
   ): Promise<EthereumTransactionTypeExtended[]> {
     const txs: EthereumTransactionTypeExtended[] = [];
@@ -587,13 +622,8 @@ export default class LendingPool
 
     const tokenToDecimals: number = await this.erc20Service.decimalsOf(toAsset);
 
-    const amountSlippage = (
-      Number(toAmount) -
-      (Number(toAmount) * Number(maxSlippage)) / 100
-    ).toString();
-
     const amountSlippageConverted: string = parseNumber(
-      amountSlippage,
+      minToAmount,
       tokenToDecimals
     );
 
@@ -601,16 +631,22 @@ export default class LendingPool
       this.lendingPoolAddress
     );
 
-    if (flash) {
-      const FLASHLOAN_PREMIUM_TOTAL: BigNumber = await lendingPoolContract.FLASHLOAN_PREMIUM_TOTAL();
-      const convertedAmountNoFees: string = BigNumber.from(convertedAmount)
-        .sub(
-          BigNumber.from(convertedAmount)
-            .mul(FLASHLOAN_PREMIUM_TOTAL)
-            .div(10000)
-        )
-        .toString();
+    const params = buildParaSwapLiquiditySwapParams(
+      toAsset,
+      amountSlippageConverted,
+      swapAll
+        ? augustusFromAmountOffsetFromCalldata(swapCallData as string)
+        : 0,
+      swapCallData,
+      augustus,
+      permitParams.amount,
+      permitParams.deadline,
+      permitParams.v,
+      permitParams.r,
+      permitParams.s
+    );
 
+    if (flash) {
       const amountWithSurplus: string = (
         Number(fromAmount) +
         (Number(fromAmount) * Number(SURPLUS)) / 100
@@ -621,38 +657,13 @@ export default class LendingPool
         tokenDecimals
       );
 
-      const params: string = utils.defaultAbiCoder.encode(
-        [
-          'address[]',
-          'uint256[]',
-          'bool[]',
-          'uint256[]',
-          'uint256[]',
-          'uint8[]',
-          'bytes32[]',
-          'bytes32[]',
-          'bool[]',
-        ],
-        [
-          [toAsset],
-          [amountSlippageConverted],
-          [swapAll],
-          [permitParams.amount],
-          [permitParams.deadline],
-          [permitParams.v],
-          [permitParams.r],
-          [permitParams.s],
-          [useEthPath || false],
-        ]
-      );
-
       const txCallback: () => Promise<transactionType> = this.generateTxCallback(
         {
           rawTxMethod: () =>
             lendingPoolContract.populateTransaction.flashLoan(
               SWAP_COLLATERAL_ADAPTER,
               [fromAsset],
-              swapAll ? [convertedAmountWithSurplus] : [convertedAmountNoFees],
+              swapAll ? [convertedAmountWithSurplus] : [convertedAmount],
               [0], // interest rate mode to NONE for flashloan to not open debt
               onBehalfOf || user,
               params,
@@ -680,12 +691,12 @@ export default class LendingPool
         user,
         assetToSwapFrom: fromAsset,
         assetToSwapTo: toAsset,
-        amountToSwap: swapAll
-          ? constants.MaxUint256.toString()
-          : convertedAmount,
+        amountToSwap: convertedAmount,
         minAmountToReceive: amountSlippageConverted,
+        swapAll,
+        swapCallData,
+        augustus,
         permitParams,
-        useEthPath,
       }
     );